pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "agents"
version = "0.1.0"
authors = [
  { name="Vishal Patel", email="vishalapatel@google.com" },
]
description = "A collection of ADK example agents."
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk==1.13.0",
    "google-cloud-aiplatform[adk,agent-engines]==1.110.0"
]

[tool.setuptools.packages.find]
where = ["agents"]